+++ example/lib/main.dart
@@ -120,8 +120,29 @@ class FrenchDepartementsList extends StatelessWidget {
   }
 }
 
-class FranceWolfMap extends StatelessWidget {
+class FranceWolfMap extends StatefulWidget {
   const FranceWolfMap({super.key});
+
+  @override
+  State<FranceWolfMap> createState() => _FranceWolfMapState();
+}
+
+class _FranceWolfMapState extends State<FranceWolfMap> {
+  final _mapKey = GlobalKey<InteractiveMapState>();
+
+  @override
+  void initState() {
+    super.initState();
+
+    Future.delayed(
+      const Duration(seconds: 1),
+      () async {
+        final countryMap =
+            await SvgParser().parse(_mapKey.currentState!.svgData!);
+        countryMap.getCountryCodeFromLocation(40.7128, -74.0060);
+      },
+    );
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -130,6 +151,7 @@ class FranceWolfMap extends StatelessWidget {
         children: [
           Flexible(
             child: InteractiveMap(
+              key: _mapKey,
               MapEntity.world,
               theme: InteractiveMapTheme(
                   borderColor: Colors.green.shade800,
@@ -166,48 +188,48 @@ class FranceWolfMap extends StatelessWidget {
               markers: [
                 MarkerGroup(
                   markers: [
-                    GeoMarker(lat: 48.864716, long: 2.349014), // Paris, France
-                    GeoMarker(
-                        lat: 38.9072, long: -77.0369), // Washington D.C., USA
-                    GeoMarker(lat: 51.5074, long: -0.1278), // London, UK
-                    GeoMarker(lat: 35.6895, long: 139.6917), // Tokyo, Japan
-                    GeoMarker(lat: -33.4489, long: -70.6693), // Santiago, Chile
-                    GeoMarker(
-                        lat: 40.7128, long: -74.0060), // New York City, USA
-                    GeoMarker(lat: 55.7558, long: 37.6176), // Moscow, Russia
-                    GeoMarker(
-                        lat: -34.6037,
-                        long: -58.3816), // Buenos Aires, Argentina
-                    GeoMarker(lat: 52.5200, long: 13.4050), // Berlin, Germany
-                    GeoMarker(lat: 40.4168, long: -3.7038), // Madrid, Spain
-                    GeoMarker(
-                        lat: -6.2088, long: 106.8456), // Jakarta, Indonesia
-                    GeoMarker(
-                        lat: -23.5505, long: -46.6333), // São Paulo, Brazil
-                    GeoMarker(
-                        lat: 55.7558, long: 12.5523), // Copenhagen, Denmark
-                    GeoMarker(lat: 59.3293, long: 18.0686), // Stockholm, Sweden
-                    GeoMarker(lat: 41.9028, long: 12.4964), // Rome, Italy
-                    GeoMarker(lat: 48.2100, long: 16.3636), // Vienna, Austria
-                    GeoMarker(lat: 45.4215, long: -75.6919), // Ottawa, Canada
-                    GeoMarker(lat: 52.2297, long: 21.0122), // Warsaw, Poland
-                    GeoMarker(
-                        lat: -22.9068,
-                        long: -43.1729), // Rio de Janeiro, Brazil
-                    GeoMarker(
-                        lat: -22.9068,
-                        long: -43.1729), // Rio de Janeiro, Brazil
-                    GeoMarker(
-                        lat: -35.2809, long: 149.1300), // Canberra, Australia
-                    GeoMarker(lat: 38.7223, long: -9.1393), // Lisbon, Portugal
-                    GeoMarker(lat: 35.6895, long: 51.3890), // Tehran, Iran
-                    GeoMarker(
-                        lat: 35.6892,
-                        long:
-                            51.3889), // Tehran, Iran (alternative coordinates)
-                    GeoMarker(lat: 41.7151, long: 44.8271), // Tbilisi, Georgia
-                    GeoMarker(lat: 32.8872, long: 13.1913), // Tripoli, Libya
-                    GeoMarker(lat: 33.8886, long: 35.4955), // Beirut, Lebanon
+                    // GeoMarker(lat: 48.864716, long: 2.349014), // Paris, France
+                    // GeoMarker(
+                    //     lat: 38.9072, long: -77.0369), // Washington D.C., USA
+                    // GeoMarker(lat: 51.5074, long: -0.1278), // London, UK
+                    // GeoMarker(lat: 35.6895, long: 139.6917), // Tokyo, Japan
+                    // GeoMarker(lat: -33.4489, long: -70.6693), // Santiago, Chile
+                    // GeoMarker(
+                    //     lat: 40.7128, long: -74.0060), // New York City, USA
+                    // GeoMarker(lat: 55.7558, long: 37.6176), // Moscow, Russia
+                    // GeoMarker(
+                    //     lat: -34.6037,
+                    //     long: -58.3816), // Buenos Aires, Argentina
+                    // GeoMarker(lat: 52.5200, long: 13.4050), // Berlin, Germany
+                    // GeoMarker(lat: 40.4168, long: -3.7038), // Madrid, Spain
+                    // GeoMarker(
+                    //     lat: -6.2088, long: 106.8456), // Jakarta, Indonesia
+                    // GeoMarker(
+                    //     lat: -23.5505, long: -46.6333), // São Paulo, Brazil
+                    // GeoMarker(
+                    //     lat: 55.7558, long: 12.5523), // Copenhagen, Denmark
+                    // GeoMarker(lat: 59.3293, long: 18.0686), // Stockholm, Sweden
+                    // GeoMarker(lat: 41.9028, long: 12.4964), // Rome, Italy
+                    // GeoMarker(lat: 48.2100, long: 16.3636), // Vienna, Austria
+                    // GeoMarker(lat: 45.4215, long: -75.6919), // Ottawa, Canada
+                    // GeoMarker(lat: 52.2297, long: 21.0122), // Warsaw, Poland
+                    // GeoMarker(
+                    //     lat: -22.9068,
+                    //     long: -43.1729), // Rio de Janeiro, Brazil
+                    // GeoMarker(
+                    //     lat: -22.9068,
+                    //     long: -43.1729), // Rio de Janeiro, Brazil
+                    // GeoMarker(
+                    //     lat: -35.2809, long: 149.1300), // Canberra, Australia
+                    // GeoMarker(lat: 38.7223, long: -9.1393), // Lisbon, Portugal
+                    // GeoMarker(lat: 35.6895, long: 51.3890), // Tehran, Iran
+                    // GeoMarker(
+                    //     lat: 35.6892,
+                    //     long:
+                    //         51.3889), // Tehran, Iran (alternative coordinates)
+                    // GeoMarker(lat: 41.7151, long: 44.8271), // Tbilisi, Georgia
+                    // GeoMarker(lat: 32.8872, long: 13.1913), // Tripoli, Libya
+                    // GeoMarker(lat: 33.8886, long: 35.4955), // Beirut, Lebanon
                   ],
                   borderColor: Colors.blue.shade700,
                   backgroundColor: Colors.blue.shade200.withOpacity(0.4),