+++ lib/interactive_country_map.dart
@@ -1,7 +1,8 @@
 library interactive_country_map;
 
-export 'src/interactive_map.dart' show InteractiveMap;
+export 'src/interactive_map.dart' show InteractiveMap, InteractiveMapState;
 export 'src/map_entity.dart' show MapEntity;
 export 'src/interactive_map_theme.dart' show InteractiveMapTheme;
 export 'src/helpers/mapping_helper.dart' show MappingHelper;
 export 'src/svg/markers.dart' show Marker, MarkerGroup, GeoMarker;
+export 'src/svg/svg_parser.dart' show SvgParser;