+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.0.5
 
 - Add loaders to load SVG from files or assets
+- Add `GeoMarker`. A marker that uses longitude and latitude coordinates
+- More Marker theming
+- Can fill the background of the map with a color
 
 ## 0.0.4